Fix DB migration 009 constraint dropping
Database migration 009 fails to run on Centos with errors
about indexes not existing. This occurs on a fresh installation
where no previous data, etc. The indexes being dropped in migration
009 were likely automatically added by some other MySQL
version/distribution so to support both cases we add a simple
check to ensure the index exists before attempting to drop it.

Change-Id: Ib6d848c1e32bbab6fdb7e30009f5c5f907fd8477
Closes-bug: #1526463