Add support for custom YAQL functions
Custom YAQL functions must be added in the 'mistral.yaql_functions'
namespace in the entry points. The entry point name is used as the
function name in YAQL. All of the Mistral YAQL functions have been
changed to a custom functions for the sake of testing.

DocImpact: The developer section in the documention has been updated.

Change-Id: I518c7a4b616089d4b70585ed85fcd81a30c93090
Implements: blueprint mistral-custom-yaql-functions

Conflicts:
	setup.cfg